@@ -6,46 +6,48 @@
   "author": "Jack Parmer <jack.parmer@proton.me>",
   "main": "electron/main.js",
   "dependencies": {
-    "@emotion/react": "^11.10.5",
-    "@mantine/core": "^5.10.4",
-    "@mantine/hooks": "^5.10.4",
-    "@tabler/icons": "^1.71.0",
+    "@emotion/react": "^11.10.8",
+    "@mantine/core": "^6.0.9",
+    "@mantine/hooks": "^6.0.9",
+    "@tabler/icons-react": "^2.17.0",
     "@tisoap/react-flow-smart-edge": "^3.0.0",
     "cors": "^2.8.5",
-    "express": "^4.17.3",
+    "express": "^4.18.2",
     "http-proxy-middleware": "^2.0.6",
-    "immer": "^9.0.19",
-    "ioredis": "^4.28.5",
-    "jotai": "^2.0.2",
+    "immer": "^10.0.1",
+    "ioredis": "^5.3.2",
+    "jotai": "^2.0.4",
     "jotai-immer": "^0.2.0",
     "js-yaml": "^4.1.0",
-    "just-clone": "^6.0.1",
+    "just-clone": "^6.2.0",
     "localforage": "^1.10.0",
     "module": "^1.2.5",
-    "plotly.js": "^2.17.0",
-    "rc-slider": "^9.7.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "plotly.js": "^2.22.0",
+    "rc-slider": "^10.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-dial-knob": "file:./bauhaus/react-dial-knob-1.3.0.tgz",
     "react-grid-layout": "^1.3.4",
-    "react-modal": "^3.14.4",
+    "react-modal": "^3.16.1",
     "react-plotly.js": "^2.6.0",
-    "react-resizable": "^3.0.4",
-    "react-scripts": "5.0.0",
-    "react-select": "^5.3.0",
+    "react-resizable": "^3.0.5",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.7.2",
     "react-switch": "^7.0.0",
     "react-syntax-highlighter": "^15.5.0",
-    "react-tabs": "^3.2.3",
+    "react-tabs": "^6.0.1",
     "react-use": "^17.4.0",
-    "reactflow": "^11.4.2",
-    "redis": "^4.0.4",
-    "socket.io": "^4.5.1",
-    "socket.io-client": "^4.5.1",
-    "styled-components": "^5.3.3",
+    "reactflow": "^11.7.0",
+    "redis": "^4.6.6",
+    "socket.io": "^4.6.1",
+    "socket.io-client": "^4.6.1",
+    "styled-components": "^5.3.10",
     "upath": "^2.0.1",
-    "use-file-picker": "^1.4.2",
-    "uuid": "^8.3.2",
+    "use-file-picker": "^1.6.1",
+    "uuid": "^9.0.0",
     "vite-plugin-environment": "^1.1.3",
-    "wait-on": "^6.0.1"
+    "wait-on": "^7.0.1",
+    "web-vitals": "^3.3.1"
   },
   "scripts": {
     "start": "vite",
@@ -124,35 +126,33 @@
   "devDependencies": {
     "@cypress/snapshot": "^2.1.7",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/react-hooks": "^8.0.1",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/file-saver": "^2.0.5",
-    "@types/node": "^18.0.0",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/node": "^18.16.3",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.1",
     "@types/react-grid-layout": "^1.3.2",
-    "@types/react-modal": "^3.13.1",
-    "@types/react-plotly.js": "^2.5.2",
+    "@types/react-modal": "^3.16.0",
+    "@types/react-plotly.js": "^2.6.0",
     "@types/styled-components": "^5.1.26",
     "@types/testing-library__jest-dom": "^5.14.5",
-    "@types/uuid": "^8.3.4",
+    "@types/uuid": "^9.0.1",
     "@vitejs/plugin-react-swc": "^3.3.0",
-    "concurrently": "^7.6.0",
-    "cypress": "^12.7.0",
+    "concurrently": "^8.0.1",
+    "cypress": "^12.11.0",
     "dotenv": "^16.0.3",
-    "electron": "^21.0.1",
-    "electron-builder": "23.0.2",
-    "eslint-plugin-cypress": "^2.12.1",
-    "jest": "^29.4.3",
-    "jest-environment-jsdom": "^29.4.3",
-    "prettier": "^2.8.8",
-    "react-dial-knob": "^1.3.0",
-    "react-error-boundary": "^4.0.1",
-    "ts-jest": "^29.0.5",
+    "electron": "^24.1.3",
+    "electron-builder": "23.6.0",
+    "eslint-plugin-cypress": "^2.13.3",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "prettier": "2.8.8",
+    "react-error-boundary": "^4.0.4",
+    "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
-    "tsconfig-paths-webpack-plugin": "^4.0.0",
-    "typescript": "^4.9.3",
+    "tsconfig-paths-webpack-plugin": "^4.0.1",
+    "typescript": "^4.9.5",
     "vite": "^4.3.4"
   }
 }
