@@ -5,15 +5,17 @@ import App from "./App";
 import { SocketContextProvider } from "./context/socket.context";
 // default styling
 import "reactflow/dist/style.css";
+import { createRoot } from "react-dom/client";
 
 // or if you just want basic styles
 import "reactflow/dist/base.css";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <React.StrictMode>
     <SocketContextProvider>
       <App />
     </SocketContextProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
