@@ -1,5 +1,5 @@
 import { ResultsType } from "@src/feature/results_panel/types/ResultsType";
-import { WritableDraft } from "immer/dist/internal";
+import { Draft } from "immer";
 import React from "react";
 import { Node, ReactFlowJsonObject } from "reactflow";
 
@@ -10,9 +10,7 @@ export interface FlowChartProps {
   setRfInstance: (
     update?:
       | ReactFlowJsonObject<any, any>
-      | ((
-          draft: WritableDraft<ReactFlowJsonObject<any, any>> | undefined
-        ) => void)
+      | ((draft: Draft<ReactFlowJsonObject<any, any>> | undefined) => void)
       | undefined
   ) => void;
 
